V0.1.0: Rewrite should auto-detect modules better
When executing `rebelle` within the folder structure of a node
module, it will attempt to load that module into the session
with the name of the module. This works by traversing the file
tree upwards until it hits a folder with a file called package.json

If a node_modules/ folder is present in the current module (or
alternatively the folder it is executed, in the case no module
has been found), it will attempt to load these modules into the
session as well.

Loading modules via the command line still works so:

    $ rebelle --hello hello.js
	> hello()
	"hello, world!"

Will still work as usual.