Bumped to v0.2.3 fixes #6
The previous implementation was kinda hackish, as it attempted to
hide the prompt ('> ') while the initialization message was
outputted by setting it to an empty string.

This fix waits till after the modules has been loaded with the
actual initialization of the prompt, and copies the loaded modules
into the session context, instead of changing the prompt to an
empty string.